Update 2019 - 07 -03
New Invoice writing to db complete
model update
db update updated sp's and new test data
invoice num generator updates

+++ TaxApp/Model/InvoiceLineItem.cs
@@ -17,7 +17,7 @@ namespace Model
         public int LineItemID { get; set; }
         public string InvoiceNum { get; set; }
         public string Name { get; set; }
-        public int UnitCount { get; set; }
+        public decimal UnitCount { get; set; }
         public decimal UnitCost { get; set; }
     }
 }
